패키지를 자바 표준 명명 규칙에 맞게 변경
- 모두 소문자로
- 복수형을 사용하지 않도록
