+++ .idea/sqldialects.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/main/java/SCP079/Commands/configCommand.java" dialect="MySQL" />
+    <file url="file://$PROJECT_DIR$/src/main/java/me/kirito5572/scp079/command/configCommand.java" dialect="MySQL" />
     <file url="PROJECT" dialect="MySQL" />
   </component>
 </project>
