+++ src/main/java/me/kirito5572/scp079/App.java
@@ -1,12 +1,12 @@
-package SCP079;/*
+package me.kirito5572.scp079;/*
  * This Java source file was generated by the Gradle 'init' task.
  */
 
-import SCP079.Listener.Listener;
-import SCP079.Listener.activityChangeListener;
-import SCP079.Listener.getHoryuBanList;
-import SCP079.Objects.CommandManager;
-import SCP079.Objects.SQLDB;
+import me.kirito5572.scp079.listener.Listener;
+import me.kirito5572.scp079.listener.activityChangeListener;
+import me.kirito5572.scp079.listener.getHoryuBanList;
+import me.kirito5572.scp079.object.CommandManager;
+import me.kirito5572.scp079.object.SQLDB;
 import me.duncte123.botcommons.messaging.EmbedUtils;
 import me.duncte123.botcommons.web.WebUtils;
 import net.dv8tion.jda.api.AccountType;
