+++ src/main/java/me/kirito5572/scp079/Constants.java
@@ -1,4 +1,4 @@
-package SCP079;
+package me.kirito5572.scp079;
 
 public class Constants {
