+++ src/main/java/me/kirito5572/scp079/command/HelpCommand.java
@@ -1,8 +1,8 @@
-package SCP079.Commands;
+package me.kirito5572.scp079.command;
 
-import SCP079.App;
-import SCP079.Objects.CommandManager;
-import SCP079.Objects.ICommand;
+import me.kirito5572.scp079.App;
+import me.kirito5572.scp079.object.CommandManager;
+import me.kirito5572.scp079.object.ICommand;
 import me.duncte123.botcommons.messaging.EmbedUtils;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
