+++ src/main/java/me/kirito5572/scp079/command/TestCommand.java
@@ -1,0 +1,27 @@
+package me.kirito5572.scp079.command;
+
+import me.kirito5572.scp079.object.ICommand;
+import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
+
+import java.util.List;
+
+public class TestCommand implements ICommand {
+    @Override
+    public void handle(List<String> args, GuildMessageReceivedEvent event) {
+    }
+
+    @Override
+    public String getHelp() {
+        return "테스트";
+    }
+
+    @Override
+    public String getInvoke() {
+        return "테스트";
+    }
+
+    @Override
+    public String getSmallHelp() {
+        return "테스트";
+    }
+}
