+++ src/main/java/me/kirito5572/scp079/object/ICommand.java
@@ -1,4 +1,4 @@
-package SCP079.Objects;
+package me.kirito5572.scp079.object;
 
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
