+++ src/main/java/me/kirito5572/scp079/object/linkConfirm.java
@@ -1,4 +1,4 @@
-package SCP079.Objects;
+package me.kirito5572.scp079.object;
 
 import java.net.HttpURLConnection;
 import java.net.URL;
